refactor mri dev && mri build && mri pro
