+++ bin/mri-build
@@ -1,15 +1,38 @@
 #!/usr/bin/env node
 
 /**
- *  mri dev ${theme}
+ *  mri build ${project}
  *  启动 MRI 项目
  */
-
 'use strict';
+const _program = require('commander');
+const $run = require('../lib/common/run');
+const $_project = require('../lib/module/project');
 
-const $_project = require('../process/project');
-
+const pk = require('../package.json');
 const MRI_ENV = 'test';
 
-$_project(MRI_ENV);
+_program
+    .usage('mri test [options]')
+    .description('运行MRI的测试环境')
+    .description('更多的环境变量\n  参考 https://umijs.org/guide/env-variables.html\n  在src/theme/${theme}/${theme}-config.js 中进行配置')
+    .option('-c, --BABEL_CACHE', '是否使用BabelCache模式')
+    .option('-n, --no-install', '忽略安装包更新')
+    .option('-i, --interface', '生成const interface文件')
+    .option('-d, --debug', 'debug 模式')
+    .option('-s, --service [name]', '服务器环境')
+    .version(pk.version);
+
+_program.parse(process.argv);
+
+$run.branch('test', () => {
+    $run.root(() => {
+        $run.package(() => {
+            $run.upgrade(() => {
+                $_project(MRI_ENV);
+            });
+        });
+    });
+});
+
