+++ bin/mri-component
@@ -13,7 +13,7 @@ const chalk = require('chalk');
 const fp = require('lodash/fp');
 const axios = require('axios');
 const cheerio = require('cheerio');
-const util = require('../service/util');
+const util = require('../lib/common/util');
 
 const DEF_OUT_PATH = {
     component: './src/components',
