+++ bin/mri-dev
@@ -6,16 +6,33 @@
  */
 
 'use strict';
+const _program = require('commander');
+const $run = require('../lib/common/run');
+const $_project = require('../lib/module/project');
 
-const $run = require('../service/run');
-const $_project = require('../process/project');
-
+const pk = require('../package.json');
 const MRI_ENV = 'dev';
 
-$run.root(() => {
-    $run.package(() => {
-        $run.upgrade(() => {
-            $run.branch('all', () => {
+_program
+    .usage('mri dev [options]')
+    .description('运行MRI的开发环境')
+    .description('更多的环境变量\n  参考 https://umijs.org/guide/env-variables.html\n  在src/theme/${theme}/${theme}-config.js 中进行配置')
+    .option('-c, --BABEL_CACHE', '是否使用BabelCache模式')
+    .option('-n, --no-install', '忽略安装包更新')
+    .option('-i, --interface', '生成const interface文件')
+    .option('-d, --debug', 'debug 模式')
+    .option('-s, --service [name]', '服务器环境')
+    .option('-f, --force', '强制启动（关闭其他node端口）')
+    .option('-p, --port [port]', 'server port')
+    .option('-b, --browser', '是否默认打开浏览器')
+    .version(pk.version);
+
+_program.parse(process.argv);
+
+$run.branch('all', () => {
+    $run.root(() => {
+        $run.package(() => {
+            $run.upgrade(() => {
                 $_project(MRI_ENV);
             });
         });
