+++ bin/mri-gco
@@ -11,7 +11,7 @@ const _fse = require('fs-extra');
 const _ = require('lodash');
 
 const $root = require('../service/root');
-const $util = require('../service/util');
+const $util = require('../lib/common/util');
 const $log = require('../service/log');
 const $template = require('../service/template');
 const $gco = require('../process/gco');
