+++ bin/mri-test
@@ -9,15 +9,15 @@ const _ = require('lodash');
 const $template = require('../service/template');
 const $root = require('../service/root');
 const $MriVersion = require('../service/mri-version-upgrade');
-const $Bus = require('../service/bus');
+const $Bus = require('../lib/services/bus');
 const _which = require('which');
 const _shell = require('shelljs');
 const _spawn = require('cross-spawn');
 const _chalk = require('chalk');
 
 const $log = require('../service/log');
-const $util = require('../service/util');
-const $run = require('../service/run');
+const $util = require('../lib/common/util');
+const $run = require('../lib/common/run');
 
 const $upgrade = require('../upgrade/v2.9.0');
