+++ service/command.js
@@ -5,7 +5,7 @@
  * 默认 < config.js 配置 < 启动命令设置的环境变量
  */
 
-const _util = require('./util');
+const _util = require('../lib/common/util');
 const _fs = require('fs');
 const _fse = require('fs-extra');
 const _ = require('lodash');
@@ -25,7 +25,7 @@ module.exports = function command(env) {
         .option('-c, --BABEL_CACHE', '是否使用BabelCache模式')
         .option('-n, --not-install', '忽略安装包更新')
         .option('-i, --INTERFACE', '生成const interface文件')
-        .option('-d, --debug', '生成const interface文件')
+        .option('-d, --debug', 'debug 模式')
         .option('-m, --mri [mri]', 'mri所在路径（no ln)')
         .option('-s, --service [name]', '服务器环境')
         .version(_pk.version);
