--- service/constant.js
@@ -1,10 +1,0 @@
-module.exports = {
-
-    // 约定根目录必须条件
-    ROOT_JUDGE: {
-        node_modules: 1,
-        'package.json': 1,
-    },
-
-    PATH_MRIRC: '.mrirc.js'
-};
