+++ service/index.js
@@ -3,7 +3,7 @@ const _join = _path.join;
 const _ = require('lodash');
 const _fs = require('fs');
 const $root = require('../service/root');
-const $util = require('../service/util');
+const $util = require('../lib/common/util');
 const _chalk = require('chalk');
 const _shell = require('shelljs');
 const _relative = require('relative');
